+++ data-structure/union_find.cpp
@@ -8,14 +8,14 @@ using namespace std;
 struct UnionFind {
     vector<int> par;
 
-    UnionFind(size_t n) : par(n, -1) {}
+    UnionFind(int n) : par(n, -1) {}
 
-    size_t find(size_t x) {
+    int find(int x) {
         if (par[x] < 0) return x;
         return par[x] = find(par[x]);
     }
 
-    void unite(size_t x, size_t y) {
+    void unite(int x, int y) {
         x = find(x);
         y = find(y);
         if (x == y) return;
@@ -24,11 +24,11 @@ struct UnionFind {
         par[y] = x;
     }
 
-    bool same(size_t x, size_t y) {
+    bool same(int x, int y) {
         return find(x) == find(y);
     }
 
-    size_t size(size_t x) {
+    int size(int x) {
         return -par[find(x)];
     }
 };
